refactor: 에러 핸들러 패키지 위치 변경 및 로그 출력 추가
