+++ src/main/java/roomescape/global/handler/GlobalExceptionHandler.java
@@ -1,0 +1,39 @@
+package roomescape.global.handler;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+    private final Logger LOGGER = LoggerFactory.getLogger(getClass());
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException exception) {
+        LOGGER.error(exception.getMessage());
+        return ResponseEntity.badRequest().body(exception.getMessage());
+    }
+
+    @ExceptionHandler(IllegalStateException.class)
+    public ResponseEntity<String> handleIllegalStateException(IllegalStateException exception) {
+        LOGGER.error(exception.getMessage());
+        return ResponseEntity.badRequest().body(exception.getMessage());
+    }
+
+    @ExceptionHandler(HttpMessageNotReadableException.class)
+    public ResponseEntity<String> handleHttpMessageNotReadableException(HttpMessageNotReadableException exception) {
+        LOGGER.error(exception.getMessage());
+        return ResponseEntity.badRequest().body("요청 형식이 올바르지 않습니다.");
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<String> handleException(Exception exception) {
+        LOGGER.error(exception.getMessage());
+        return ResponseEntity.internalServerError().body("서버 내부에서 오류가 발생했습니다.");
+    }
+
+}
