--- src/main/java/roomescape/handler/GlobalExceptionHandler.java
@@ -1,32 +1,0 @@
-package roomescape.handler;
-
-import com.fasterxml.jackson.databind.exc.InvalidFormatException;
-import org.apache.coyote.Response;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException exception){
-        return ResponseEntity.badRequest().body(exception.getMessage());
-    }
-
-    @ExceptionHandler(IllegalStateException.class)
-    public ResponseEntity<String> handleIllegalStateException(IllegalStateException exception){
-        return ResponseEntity.badRequest().body(exception.getMessage());
-    }
-
-    @ExceptionHandler(InvalidFormatException.class)
-    public ResponseEntity<String> handleInvalidFormatException(InvalidFormatException exception){
-        return ResponseEntity.badRequest().body("유효하지 않은 형식입니다.: " + exception.getMessage());
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleException(Exception exception){
-        return ResponseEntity.internalServerError().body("서버 내부 오류입니다.: " + exception.getMessage());
-    }
-
-}
